+++ yamlator/parser/loaders.py
@@ -333,8 +333,10 @@ def resolve_ruleset_inheritance(rulesets: Dict[str, YamlatorRuleset]) -> dict:
     Raises:
         ValueError: If the ruleset parameter is `None`
         TypeError: If the ruleset parameter is not a `dict`
-        yamlator.exceptions.ConstructNotFoundError: If the parent of the ruleset
-            cannot be found in the `rulesets` parameter
+        yamlator.exceptions.ConstructNotFoundError: If a parent ruleset
+            that is being inherited cannot be found in the `rulesets` parameter
+        yamlator.exceptions.CycleDependencyError: If there is a cycle in the
+            inheritance chain
     """
     if rulesets is None:
         raise ValueError('Parameter rulesets cannot be None')
@@ -345,31 +347,66 @@ def resolve_ruleset_inheritance(rulesets: Dict[str, YamlatorRuleset]) -> dict:
 
     updated_rulesets = {}
 
+    dependencies_mgmr = DependencyManager()
     for key, ruleset in rulesets.items():
-        if ruleset.parent is None:
+        parent = ruleset.parent
+        if not parent:
             updated_rulesets[key] = ruleset
             continue
 
-        parent_name = ruleset.parent.lookup
-        parent = rulesets.get(parent_name)
-        if parent is None:
+        if rulesets.get(parent.lookup) is None:
             raise ConstructNotFoundError(parent)
 
-        base_rules = ruleset.rules.copy()
-        parent_rules = parent.rules.copy()
+        dependencies_mgmr.add_child(key, ruleset.parent.lookup)
 
-        # Index the rules in the base and parent rulesets to make it
-        # easier to merge the different rules together
-        base_rules_index = {rule.name: rule for rule in base_rules}
-        parent_rules_index = {rule.name: rule for rule in parent_rules}
+    if dependencies_mgmr.has_cycle():
+        msg = 'Detected cycle when resolving inheritance chain'
+        raise CycleDependencyError(msg)
 
-        # Merged the 2 rule lists together. If a rule name is present
-        # in both the base rules will be prioritised since it assumed
-        # it is being overridden
-        merged_rules = list({**parent_rules_index, **base_rules_index}.values())
-        updated_rulesets[key] = YamlatorRuleset(
-            name=ruleset.name,
-            rules=merged_rules,
-            is_strict=ruleset.is_strict
-        )
+    dependencies = dependencies_mgmr.graph
+    for node in dependencies:
+        if updated_rulesets.get(node) is not None:
+            continue
+
+        stack = [(node, None)]
+        while len(stack) > 0:
+            curr_node, curr_ruleset = stack.pop()
+
+            if curr_ruleset is not None:
+                parent_node, _ = stack.pop()
+                parent_ruleset = rulesets[parent_node]
+                updated_rulesets[parent_node] = _merge_rulesets(parent_ruleset,
+                                                                curr_ruleset)
+                if len(stack) > 1:
+                    stack.append((parent_node, updated_rulesets[parent_node]))
+                continue
+
+            if updated_rulesets.get(curr_node) is not None:
+                stack.append((curr_node, updated_rulesets[curr_node]))
+                continue
+
+            stack.append((curr_node, None))
+            stack.append((dependencies[curr_node][0], None))
+
     return updated_rulesets
+
+
+def _merge_rulesets(ruleset: YamlatorRuleset,
+                    dependent_ruleset: YamlatorRuleset) -> YamlatorRuleset:
+    base_rules = ruleset.rules.copy()
+    dependent_rules = dependent_ruleset.rules.copy()
+
+    # Index the rules in the base and dependent rulesets to make it
+    # easier to merge the different rules together
+    base_rules_index = {rule.name: rule for rule in base_rules}
+    dependent_rules_index = {rule.name: rule for rule in dependent_rules}
+
+    # Merged the 2 rule lists together. If a rule name is present
+    # in both then the base rules will be prioritized since it assumed
+    # it is being overridden
+    merged_rules = list({**dependent_rules_index, **base_rules_index}.values())
+    return YamlatorRuleset(
+        name=ruleset.name,
+        rules=merged_rules,
+        is_strict=ruleset.is_strict
+    )